+++ src/components/common/Modal/LoginModal/LoginModal.module.scss
@@ -1,0 +1,52 @@
+.modalBackground {
+  padding: 4px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  color: var(--lukyviky-Blue-Darker, #002b59);
+  font-size: 26px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 30px;
+  letter-spacing: 1.04px;
+  height: 226px;
+  gap: 6px;
+
+  @media only screen and (max-width: 520px) {
+    font-size: 16px;
+  }
+
+  .title {
+    color: var(--lukyviky-Blue-Normal, #007aff);
+    text-align: center;
+    font-size: 26px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 30px; /* 115.385% */
+    letter-spacing: 1.04px;
+    @media only screen and (max-width: 520px) {
+      font-size: 20px;
+    }
+  }
+}
+
+.modalFullButton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--lukyviky-Blue-Light, #e6f2ff);
+  background: var(--lukyviky-Blue-Normal, #007aff);
+  display: flex;
+  width: 100%;
+  height: 113px;
+  font-size: 24px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 18px; /* 75% */
+  letter-spacing: 0.96px;
+  text-align: center;
+  @media only screen and (max-width: 520px) {
+    font-size: 16px;
+  }
+}
